refactor combat tab into smaller components
