+++ components/combat/combatStories.helpers.ts
@@ -1,0 +1,73 @@
+import type { Character } from "@/lib/character-types"
+
+export const basicCharacter: Character = {
+  id: "1",
+  name: "Test",
+  attributes: {
+    fortitude: { base: 1, added: 0, bonus: 0 },
+    finesse: { base: 1, added: 0, bonus: 0 },
+    force: { base: 1, added: 0, bonus: 0 },
+  },
+  abilities: {
+    athletics: { base: 0, added: 0, bonus: 0 },
+    awareness: { base: 0, added: 0, bonus: 0 },
+    closeCombat: { base: 2, added: 0, bonus: 0 },
+    craft: { base: 0, added: 0, bonus: 0 },
+    embassy: { base: 0, added: 0, bonus: 0 },
+    integrity: { base: 0, added: 0, bonus: 0 },
+    navigate: { base: 0, added: 0, bonus: 0 },
+    physique: { base: 0, added: 0, bonus: 0 },
+    presence: { base: 0, added: 0, bonus: 0 },
+    performance: { base: 0, added: 0, bonus: 0 },
+    rangedCombat: { base: 1, added: 0, bonus: 0 },
+    sagacity: { base: 0, added: 0, bonus: 0 },
+    stealth: { base: 0, added: 0, bonus: 0 },
+    war: { base: 0, added: 0, bonus: 0 },
+  },
+  essence: { rating: 1, motes: 5, commitments: 0, spent: 0, anima: 0 },
+  staticValues: {
+    defenseModifier: 0,
+    evasionModifier: 0,
+    parryModifier: 0,
+    resolveModifier: 0,
+    soakModifier: 0,
+    hardnessModifier: 0,
+  },
+  health: {
+    baseline: { zero: 1, minusOne: 2, minusTwo: 2, incap: 1 },
+    oxBodyLevels: 0,
+    exaltType: "lunar",
+    bashingDamage: 0,
+    lethalDamage: 0,
+    aggravatedDamage: 0,
+    dramaticInjuries: [],
+  },
+  armor: [],
+  weapons: [],
+  milestones: { personal: 0, exalt: 0, minor: 0, major: 0 },
+  advancement: [],
+  dicePool: {
+    attribute: "fortitude",
+    ability: "athletics",
+    targetNumber: 7,
+    doublesThreshold: 10,
+    extraDiceBonus: 0,
+    extraDiceNonBonus: 0,
+    extraSuccessBonus: 0,
+    extraSuccessNonBonus: 0,
+    isStunted: false,
+  },
+  charms: [],
+  spells: [],
+  combat: {
+    power: 0,
+    joinBattleBonus: 0,
+    joinBattleDiceBonus: 0,
+    joinBattleSuccessBonus: 0,
+    decisiveExtraDice: 0,
+    decisiveExtraSuccess: 0,
+  },
+  social: { virtues: { major: null, minor: null }, intimacies: [], backgrounds: [] },
+  rulings: [],
+}
+
